requirements.txt
openai
elasticsearch==8.10.1
python-dotenv
folium
